Trading with AI/01-Simple Data Streaming/requirements.txt
kafka-python
websocket_client